@@ -73,8 +73,8 @@
     "p-map": "^4.0.0",
     "patch-package": "^6.2.1",
     "pify": "^5.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^11.3.3",
     "react-icons": "^4.1.0",
     "react-lottie-player": "^1.3.3",
