@@ -1,5 +1,5 @@
 import React, { Suspense } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import App from './App';
 import ErrorBoundary from './ErrorBoundary';
@@ -14,12 +14,12 @@ const { app } = window.require('@electron/remote');
 console.log('Version', app.getVersion());
 
 
-ReactDOM.render(
-  (
-    <ErrorBoundary>
-      <Suspense fallback={<div />}>
-        <App />
-      </Suspense>
-    </ErrorBoundary>
-  ), document.getElementById('root'),
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
+  <ErrorBoundary>
+    <Suspense fallback={<div />}>
+      <App />
+    </Suspense>
+  </ErrorBoundary>,
 );
